drm/shmem-helper: Make pages_use_count an atomic_t
This way we can grab a pages ref without acquiring the resv lock when
pages_use_count > 0. This is needed to implement asynchronous map using
the drm_gpuva_mgr when the map/unmap operation triggers a mapping split,
requiring the new left/right regions to grab an additional page ref
to guarantee that the pages stay pinned when the middle section is
unmapped.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Steven Price <steven.price@arm.com>